adds verification for reset_db endpoint

+++ project/blueprints/reset_blueprint.py
@@ -1,6 +1,9 @@
-from flask_openapi3 import APIBlueprint
+from flask_openapi3 import APIBlueprint, Info
 from project import db
 from flask import jsonify
+from decouple import config
+
+info = Info(title='Reset endpoint', description="Resets the DB, must be used only for testing purposes")
 
 RESET_ENDPOINT = "/database/reset"
 
@@ -12,6 +15,11 @@ reset_blueprint = APIBlueprint("reset_blueprint", __name__)
     f"{RESET_ENDPOINT}"
 )
 def reset_database():
-    db.drop_all()
-    db.create_all()
-    return jsonify({"message": "Database reset"})
+    reset_allowed = config("ALLOW_DB_RESET", cast=bool, default=False)
+
+    if reset_allowed:
+        db.drop_all()
+        db.create_all()
+        return jsonify({"message": "Database reset"})
+    else:
+        return jsonify({"message": "Database reset not allowed"}), 403
